Move prepare_to_upload to infractl.fs module
